More edits to Chpt 1
Add the Evil static global var, with appriopriate comments.  Drop the singleton class, and every place it had to be passed.  Much smaller code now, and will be a LOT smaller as the project grows.

Drop the opcode; minor usages in prior implementations for performance, not needed here.

Use the preview pattern match in test code to simplify.