Replace Parser scopes with ScopeNode
Basically move the Stack of HashMaps to a ScopeNode, with wrappers.
The HashMap maps String names to Integer indices; inputs to the ScopeNode.
This way, the ScopeNode keeps alive Nodes that are only alive because they are alive in the HashMap.

Fix bug isDead -> isUnused in deadCodeElim.
Support routines for adding/removing edges in ScopeNode.
Set $ctrl to null after a Return; there is no control.
Regression tests.