Bugfix 78 (#81)
* Missing idepth for LoopNode

* backedge code removing a Parser keep

Stop subsuming a node parser yet needs

* Missing addDep

* LoadNode cloning peep better profit

The LoadNode cloning peep cannot clone around a Loop unless its profit on the backedge, otherwise might infinitely clone around backedge.
peep: x - (-y) is x+y
peep: -(-x) is x

* Drop alias from Field

Field is intended to be a part of a Type; as such is abstract, immutable, and true for ALL Time and all TypeStructs which have the same field.  Drop the alias and struct from Field.  Drop field reset.
The alias#s are decided in the Parser, and have a lifetime for the compilation unit; the Fields have infinite lifetime.
Minor bugfix idepth caching

* Remove emacs junk file