Harden scheduler against infinite loops
I hardened the scheduler a bunch against this, but the results are mostly nonsensical in any case.
I had to add a Constant 0 in the forced exit path... which blows the CME assert in ArrayList on Start when a new constant is created, so then I pre-cooked a constant 0 in the parser so I wouldn't hit the CME, which renumbers all the golden tests because the Node numbers shift by one.
:-P