Change field mutability
Fields & locals with an initializer are immutable by default; those without are mutable.  Make mutable with `!`.  Fields are always mutable in a constructor.
Add a trinary `pred ? then : else` syntax.
SubNode retains as much precision as AddNode.
Sharpen some asserts and error cases.
Limit Load node cloning if the ptr is not obvious.