Some dead inf call bug fixes
More comments about meeting nil above the center line falls - but does not (can not) dual.
This implies have TFPs with *no* allowed functions that are below centerline; if allowed null then its basically an alternate null.  If not allowed null and not allowed any functions, its basically a high value (not executable).  IfNode folds.

Do not remove a dead infinite loop - it needs to remain as a CFG sink.