More RegAlloc
Forcing kills on allocations broke a bunch of things.
Color: pickRisky single-Def cloneables that are not next to their uses.  I.e., a X86 flags-producer far away from a required jump:
Pick as Risky, fail to color, and spill (clone) next to jump.
Remove a bunch of premature split fold-thrus