Hitting self-recursive types
Changes for allowing locals to be non-const by default.
Changes to track `_one` single-ton fields that will get moved to a class.
Field offset constants; will become a real int constant once layout is determined, which waits until fields are decided to be in the class or instance; also if the field contents can be put in a smaller size.